+++ src/app/page.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { ModeToggle } from "@/components/themes/theme-mode";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { useUtils } from "@/lib/utils";
@@ -16,11 +15,10 @@ export default function Home() {
         </div>
         <Input placeholder="Email" type="email" />
         <Input placeholder="Senha" type="password" />
-        <Button onClick={() => router.push("/dashboard")}>
+        <Button onClick={() => router.push("/pages/dashboard")}>
           <ArrowTopRightIcon />
           Entrar
         </Button>
-        <ModeToggle />
       </main>
     </div>
   );
