+++ src/containers/container-page.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+interface ContainerPageProps {
+  children: React.ReactNode;
+}
+
+const ContainerPage: React.FC<ContainerPageProps> = ({ children }) => {
+  return <div className="p-4 h-full">{children}</div>;
+};
+
+export default ContainerPage;
